Suppress warning messages about ZALLOC
This patch will suppress following warning:

```
/home/watson/.rbenv/versions/3.4.8/include/ruby-3.4.0/ruby/internal/memory.h:234:49: warning: argument 1 range [18446744071562067968, 18446744073709551615] exceeds maximum object size 9223372036854775807 [-Walloc-size-larger-than=]
  234 | #define RB_ZALLOC_N(type,n) RBIMPL_CAST((type *)ruby_xcalloc((n), sizeof(type)))
      |                                                 ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/home/watson/.rbenv/versions/3.4.8/include/ruby-3.4.0/ruby/internal/cast.h:31:29: note: in definition of macro ‘RBIMPL_CAST’
   31 | # define RBIMPL_CAST(expr) (expr)
      |                             ^~~~
/home/watson/.rbenv/versions/3.4.8/include/ruby-3.4.0/ruby/internal/memory.h:401:20: note: in expansion of macro ‘RB_ZALLOC_N’
  401 | #define ZALLOC_N   RB_ZALLOC_N   /**< @old{RB_ZALLOC_N} */
      |                    ^~~~~~~~~~~
../../../../ext/cumo/narray/index.c:827:19: note: in expansion of macro ‘ZALLOC_N’
  827 |     na2->stridx = ZALLOC_N(cumo_stridx_t,ndim_new);
      |                   ^~~~~~~~
In file included from /home/watson/.rbenv/versions/3.4.8/include/ruby-3.4.0/ruby/defines.h:71:
/home/watson/.rbenv/versions/3.4.8/include/ruby-3.4.0/ruby/internal/xmalloc.h:147:7: note: in a call to allocation function ‘ruby_xcalloc’ declared here
  147 | void *ruby_xcalloc(size_t nelems, size_t elemsiz)
      |       ^~~~~~~~~~~~
```

+++ ext/cumo/narray/index.c
@@ -799,7 +799,7 @@ VALUE cumo_na_aref_md_protected(VALUE data_value)
     int keep_dim = data->keep_dim;
     int at_mode = data->at_mode;
 
-    int ndim_new;
+    size_t ndim_new;
     VALUE view;
     cumo_narray_view_t *na2;
     ssize_t elmsz;
